requirements.txt
selenium>=4.20.0
pandas==2.2.1
openpyxl==3.1.2
python-dateutil==2.9.0